Move SortControl into components folder
