+++ src/pages/BorrowPage/components/SortNfts/SortNfts.tsx
@@ -4,7 +4,7 @@ import cx from 'classnames';
 import { useOnClickOutside } from '@frakt/hooks';
 import { BorrowNft } from '@frakt/api/nft';
 import { SearchInput } from '../../../../components/SearchInput';
-import SortControl from '../../../../componentsNew/SortControl';
+import SortControl from '../../../../components/SortControl';
 import Button from '../../../../components/Button';
 import styles from './SortNfts.module.scss';
 import {
